Heading size changes
This reduces the size of the H1 and H2 elements to avoid them dominating
real page layouts and bring them more inline with the original design of
the product.

As part of this I have also fixed an issue with a missing Babel plugin,
hence the other changes to the build tools.